Core/Movement: Minor refactors in WaypointMovementGenerator
* Removed unneccessary _loadedFromDB member (handled by variant index already)
* Apply some const
* Replace stringstream formatting in GetDebugInfo with Trinity::StringFormat
* Moved MovementGeneratorCreator::Create implementations to source files whose movement they create
